# `qmk find`, `qmk mass-compile` search speed improvements