## Description

`qmk find` and `qmk mass-compile` have been quite slow, mainly due to misuse of `lru_cache` as well as logically doing things more often than necessary.

This PR introduces the following changes:
* fixing up usage of `@lru_cache` so that it actually functions as intended
* a new `@profile_code` decorator, allowing executions call tracing to be output (requires `yappi` to be installed by `pip`)
* deferring validation of the resulting json when necessary
* a fix to `data/mappings/keyboard_aliases.hjson`
* optional de-parallelisation of `qmk.search.search_keymap_targets`, which is necessary for code profiling to make sense (profiling does not like parallel execution)

## Types of Changes

- [x] Core
- [x] Bugfix
- [x] New feature
- [x] Enhancement/optimization
- [ ] Keyboard (addition or update)
- [ ] Keymap/layout/userspace (addition or update)
- [ ] Documentation

## Checklist

- [x] My code follows the code style of this project: [**C**](https://docs.qmk.fm/#/coding_conventions_c), [**Python**](https://docs.qmk.fm/#/coding_conventions_python)
- [x] I have read the [**PR Checklist** document](https://docs.qmk.fm/#/pr_checklist) and have made the appropriate changes.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the [**CONTRIBUTING** document](https://docs.qmk.fm/#/contributing).
- [ ] I have added tests to cover my changes.
- [x] I have tested the changes and verified that they work and don't break anything (as well as I can manage).