@tzarc 2023-07-16
Performance comparison for `time qmk find -f 'split.enabled=true'` --

current `master`:
```
% time qmk find -f 'split.enabled=true
<snip>
387.62s user 4.72s system 880% cpu 44.582 total
```

this PR:
```
% time qmk find -f 'split.enabled=true
<snip>
119.02s user 2.04s system 586% cpu 20.624 total
```